refactor(storage): move channel adapters into the adapters package
Group the channel-to-DataInput/ImageInputStream adapters with the
RangeReader-to-channel adapters in io.tileverse.storage.adapters.
The full bridge from storage to standard Java IO now lives in one
package, and io.tileverse.io keeps only IO primitives such as the
buffer pool and byte ranges.

Also re-package the adapter tests out of the stale
io.tileverse.storage.rangereader.adapters package left over from
the 2.0 rename.

on-behalf-of: @camptocamp <info@camptocamp.com>
