+++ tileverse-storage/core/src/main/java/io/tileverse/storage/adapters/ReadableByteChannelDataInput.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.io;
+package io.tileverse.storage.adapters;
 
 import java.io.DataInput;
 import java.io.EOFException;
