+++ tileverse-storage/core/src/main/java/io/tileverse/storage/adapters/SeekableByteChannelImageInputStream.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.io;
+package io.tileverse.storage.adapters;
 
 import java.io.IOException;
 import java.nio.ByteBuffer;
