+++ tileverse-storage/core/src/test/java/io/tileverse/storage/adapters/RangeReaderSeekableByteChannelIntegrationTest.java
@@ -13,13 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.storage.rangereader.adapters;
+package io.tileverse.storage.adapters;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import io.tileverse.storage.RangeReader;
 import io.tileverse.storage.RangeReaderTestSupport;
-import io.tileverse.storage.adapters.RangeReaderSeekableByteChannel;
 import java.io.IOException;
 import java.nio.ByteBuffer;
 import java.nio.channels.SeekableByteChannel;
