+++ tileverse-storage/core/src/test/java/io/tileverse/storage/adapters/RangeReaderSeekableByteChannelTest.java
@@ -13,14 +13,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.storage.rangereader.adapters;
+package io.tileverse.storage.adapters;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import io.tileverse.storage.RangeReader;
 import io.tileverse.storage.RangeReaderTestSupport;
-import io.tileverse.storage.adapters.RangeReaderSeekableByteChannel;
 import io.tileverse.storage.it.TestUtil;
 import java.io.IOException;
 import java.nio.ByteBuffer;
