+++ tileverse-storage/core/src/test/java/io/tileverse/storage/adapters/ReadableByteChannelDataInputTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.tileverse.io;
+package io.tileverse.storage.adapters;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
